pyproject.toml
[tool.poetry]
name = "publixsubsale"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
selenium = "^4.12.0"
loguru = "^0.7.0"
faunadb = "^4.5.1"
python-dotenv = "^1.0.0"
firefish-py = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
